+++ lib/test/clojure_lsp/queries_test.clj
@@ -650,17 +650,35 @@
       {:ns nil, :name "unregistered-kw", :filename (h/file-path "/aaa.clj"), :bucket :keyword-usages}
       (q/find-definition-from-cursor db (h/file-path "/aaa.clj") unreg-kw-r unreg-kw-c))))
 
-;; Uncoment after clj-kondo solves https://github.com/clj-kondo/clj-kondo/issues/1632
-#_(deftest find-definition-form-java-class-usage
-    (h/load-code-and-locs (h/code "package project;"
-                                  "class Foo {}") (h/file-uri "file:///project/Foo.java"))
-    (h/load-code-and-locs (h/code "123456") (h/file-uri "file:///project/Foo.class"))
-    (testing "Finding java source even if class exists"
-      (let [[[foo-r foo-c]] (h/load-code-and-locs (h/code "(ns a (:import (project Foo)))"
-                                                          "(|Foo.)") (h/file-uri "file:///a.clj"))]
-        (h/assert-submap
-          {} ;; TODO
-          (q/find-definition-from-cursor @db/db (h/file-path "/a.clj") foo-r foo-c @db/db)))))
+(deftest find-clojure-definition-of-imported-java-class-usage
+  (h/load-code-and-locs (h/code "(ns my.fabulous-namespace)"
+                                "(defrecord SomeRecord [foo bar])"
+                                "(defprotocol SomeProtocol"
+                                "  (do-stuff [this]))") (h/file-uri "file:///project/my/fabulous_namespace.clj"))
+
+  (testing "Finding defrecord definition from imported java class"
+    (let [[[from-ns-row from-ns-col]
+           [from-defrecord-row from-defrecord-col]]
+          (h/load-code-and-locs
+            (h/code "(ns my.other-namespace"
+                    "  (:require [my.fabulous-namespace :as foobar])"
+                    "  (:import [my.fabulous_namespace |SomeRecord]))"
+                    ""
+                    "(extend-type |SomeRecord"
+                    "  foobar/SomeProtocol "
+                    "  (do-stuff [this]"
+                    "    42))") (h/file-uri "file:///project/my/other_namespace.clj"))
+          expected {:ns 'my.fabulous-namespace
+                    :name 'SomeRecord
+                    :defined-by 'clojure.core/defrecord
+                    :filename "/project/my/fabulous_namespace.clj"
+                    :bucket :var-definitions}]
+      (h/assert-submap
+        expected
+        (q/find-definition-from-cursor (h/db) (h/file-path "/project/my/other_namespace.clj") from-ns-row from-ns-col))
+      (h/assert-submap
+        expected
+        (q/find-definition-from-cursor (h/db) (h/file-path "/project/my/other_namespace.clj") from-defrecord-row from-defrecord-col)))))
 
 (deftest find-declaration-from-cursor
   (h/load-code-and-locs (h/code "(ns foo.baz) (def other 123)"))