+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   - Support java class decompilation for zipfile scheme (vim users).
   - Avoid keeping diagnostics of external closed files for Calva. https://github.com/BetterThanTomorrow/calva/issues/1864
   - Lint opened files after a clojure-lsp or clj-kondo config file is saved on disk, avoiding users to re-edit files. #1247
+  - Allow find definition of java class usages where definition comes from clojure, like defrecords.
 
 - API/CLI
   - Fix missing diagnostics when `--project-root` is different than current directory. #1245